refactor: 컨트롤러가 AttendanceFileReader를 주입받도록 변경
